fix a couple of GC issues
The first issue was that references metadata objects (e.g. Unicode tables) were
causing GC to out-of-bounds trap or infinite loop.  I addressed this by
considering any objects outside the heap as "already marked".

The second issue was that GC.resizeHeapConsistent assumed that shrinking the
heap actually shrinks the heap, whereas WasmHeap.resizeHeap does nothing when
asked to shrink.  This lead to inconsistent state in GC and an eventual
out-of-bounds trap.  I address this by double-checking that the heap really
shrunk in GC.resizeHeapConsistent and doing nothing if not.  This caused the
heap to expand more quickly than it otherwise would, so I slowed that down by
removing GC.youngGCCount and its associated logic.

There are probably better solutions to both issues, but these work well enough
for now.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>